requirements.txt
# --- Framework & Server ---
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
flask-bcrypt>=1.0.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0

# --- LangChain Ecosystem (Micro) ---
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.2
# langchain-community # REMOVED: Too large (50MB+ unzipped)
# langchain # REMOVED: Too large. We use custom agent loop.

# --- Vector Database (DATASTAX ASTRA DB) ---
langchain-astradb>=0.1.0
astrapy>=0.7.7

# --- Reranking & AI ---
openai>=1.10.0
tiktoken>=0.5.2

# --- PDF Processing ---
pypdf>=4.0.1 

# --- Frontend ---
# streamlit>=1.30.0 # NOT NEEDED FOR FLASK DEPLOYMENT

byteplus-python-sdk-v2
python-dotenv
ddgs
vercel-blob
gTTS>=2.5.1
langgraph
langsmith
langchain-groq
faiss-cpu
